Fix module bootstrap for soft navigation
